Add a Drink Feature: Ingredients
Now able to add/remove ingredients when creating a new drink. Did some minor refactoring as well and small code cleanup. Painful learning point of this fix was binding an array of strings to ngModel inside ngFor does not mix well. Had to refactor API service to use a new Ingredient[] object as opposed to a String[]. Not typically ideal to change the API service but maybe it's code smelly anyway that the rest service was returning a String[] as opposed to something else. In either case, there was little to no support online at the time for handling binding a String[] to ngModel.